#3476: created internal ExpressionCombiner class
